refactor: decompose read.go into xref.go
Extract cross-reference parsing into xref.go: xref struct,
readXref, readXrefTable, readXrefTableData, readXrefStream,
readXrefStreamData, processXrefEntry, parseWArray, decodeInt,
findLastLine, findStartxrefFallback, findStartxrefOffset,
validatePDFHeader.

read.go retains Reader lifecycle, object resolution, and filter
decoding only.
